+++ src/core/component/functional/const.ts
@@ -17,25 +17,7 @@ export const componentOpts = [
 	'components'
 ];
 
-export const destroyCheckHooks = [
-	'beforeMounted',
-	'created',
-	'beforeDestroy'
-];
-
 export const destroyHooks = [
 	'beforeDestroy',
 	'destroyed'
 ];
-
-export const mountHooks = [
-	'beforeMounted',
-	'beforeUpdated',
-	'beforeActivated'
-];
-
-export const parentMountMap = Object.createDict({
-	beforeMount: 'beforeMounted',
-	beforeUpdate: 'beforeUpdated',
-	deactivated: 'beforeActivated'
-});